bdev/nvme: Change nvme_async_probe_ctx->count type to uint32_t
This also requires change of type for count field
in rpc_bdev_nvme_attach_controller_ctx structure
and argument type in spdk_bdev_nvme_create function.


Reviewed-on: https://review.gerrithub.io/c/spdk/spdk/+/470135
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>

+++ module/bdev/nvme/common.h
@@ -87,7 +87,7 @@ struct nvme_async_probe_ctx {
 	struct spdk_nvme_probe_ctx *probe_ctx;
 	const char *base_name;
 	const char **names;
-	size_t *count;
+	uint32_t *count;
 	uint32_t prchk_flags;
 	struct spdk_poller *poller;
 	struct spdk_nvme_transport_id trid;
